[#68] fix styled-compoenets setting
